refactor(network): move to converters-protobuf::SignedBlockHeader to network SignedBlockHeader
